Add glossary entry point script and bundle
