+++ frontend/main.jsx
@@ -7,27 +7,6 @@ import { Provider } from 'react-redux';
 import { Router, browserHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
 
-import Glossary from 'glossary-panel';
-
-const terms = [
-  {
-    term: "Federalist",
-    definition: "A cool app that does some cool stuff and is cool",
-  },
-  {
-    term: "Mammal",
-    definition: 'warm blooded animal with hair or fur that gives birth to live young',
-  },
-  {
-    term: 'Platypus',
-    definition: 'A mammal that breaks all the rules',
-  },
-];
-
-new Glossary(terms, {}, {
-  termClass: 'glossary__term accordion__button',
-});
-
 import routes from './routes';
 import store from './store';
