+++ views/base.njk
@@ -126,6 +126,10 @@
             </ul>
           </nav>
         </div>
+        <div class="usa-grid">
+          <button class="js-glossary-toggle">Glossary</button>
+        </div>
+        {% include "glossary.njk" %}
       </div>
       <div class="usa-footer-secondary_section">
         <div class="usa-grid">
@@ -150,5 +154,8 @@
         </div>
       </div>
     </footer>
+    {% if not isAuthenticated %}
+      <script src="/{{ webpackAssets['glossary.js'] }}"></script>
+    {% endif %}
   </body>
 </html>
